requirements.txt
django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
python-decouple==3.8
better-sqlite3-driver==1.0.0
django-filter==24.2
whitenoise==6.6.0
gunicorn==21.2.0
